run: added support for selecting an image to run via docker image ID
This commit modifies the store's schema to store the docker id for any
image imported into the store. This value comes from the
"appc.io/docker/imageid" annotation on the image being imported. If this
annotation is missing, the stored dockerid is blank.

When an image is being run, if the schema is "docker-hash", rkt will
search for an image with this docker image ID in the store (using the
field that was added for quick lookup).