run/prepare/fetch: replace --no-store and --store-only with --pull-policy
This replaces the `--no-store` and `--store-only` flags with a singular
flag `--pull-policy` that can accept one of three things, `never`,
`new`, and `update`. This is purely a cli change, there are no under the
hood impacts. `--no-store` has been aliased to `--pull-policy=update`
and `--store-only` has been aliased to `--pull-policy=never` for
compatibility reasons.